adding basic support for soc crosswalking
